+++ object-stringifier.js
@@ -26,10 +26,6 @@ class ObjectStringifier extends Stringifier {
 	decl(node, semicolon) {
 		let prop = this.rawValue(node, 'prop');
 
-		if (prop === 'float') {
-			prop = 'cssFloat';
-		}
-
 		let string = prop;
 
 		const isObjectShorthand = node.raws.node && node.raws.node.shorthand;
@@ -100,7 +96,7 @@ class ObjectStringifier extends Stringifier {
 			!/:/.test(value)
 		) {
 			value = `:${value}`;
-		} else if (own === 'before' && /^(decl|rule)$/.test(node.type)) {
+		} else if (own === 'before' && /^(?:decl|rule)$/.test(node.type)) {
 			value = value.replace(/\S+$/, '');
 		}
 
@@ -123,11 +119,12 @@ class ObjectStringifier extends Stringifier {
 			return value;
 		}
 
-		if (/^(prop|selector)$/i.test(prop)) {
+		if (/^(?:prop|selector)$/i.test(prop)) {
 			value = camelCase(value);
 
+			// eslint-disable-next-line regexp/no-unused-capturing-group -- TODO: fix
 			if (node.raws.before && /(\S+)$/.test(node.raws.before)) {
-				value = RegExp.$1 + value;
+				value = RegExp.$1 + value; // eslint-disable-line regexp/no-legacy-features -- TODO: fix
 			} else if (value && !/\W/.test(value)) {
 				return value;
 			}